Moved logging to its own module for clarity
