+++ app/app.ts
@@ -7,6 +7,7 @@ const dotenv = require('dotenv').config();
 
 import { Request, Response } from "express-serve-static-core";
 import { Express } from "express";
+import * as Logger from "./logging/logger";
 // import * as handler from "./handler";
 import * as userHandler from "./handlers/userHandler";
 import fs from 'fs';
@@ -43,160 +44,27 @@ updateServiceAccountWithSecrets();
 let db = admin.firestore();
 // -----------------------------------------------------------------------------
 
-let tabSpaceSize = 2;
-let tabStr = "  ";
-
 function shell(thisArg: any, f: Function, req: Request, res: Response, args?: any[]) {
-  requestBoxLog("REQUEST TO: " + req.url);
-  var ip = req.headers['x-forwarded-for'] ||
-    req.connection.remoteAddress ||
-    req.socket.remoteAddress ||
-    ((req.connection as any)['socket'] ? (req.connection as any)['socket'].remoteAddress : null);
-  console.log(tabStr + "├─── BY: " + ip);
-  console.log(tabStr + "├─── AT: " + new Date().toString());
   if (req.get('eve-pk') != process.env.EVE_PK) {
-    console.log(tabStr + "├─── STATUS: REJECTED");
-    console.log(tabStr + (req.url.toLowerCase() != "/logs/" ? "├" : "└") + "─── CALLED: " + f.name);
     let eve_pk_json = { "error": "The header \"eve-pk\" was undefined or incorrect. To obtain the pk value, consult a TPM or dev lead." };
     res.status(401).json(eve_pk_json);
     return;
   } else {
-    console.log(tabStr + "├─── STATUS: ACCEPTED");
-    console.log(tabStr + (req.url.toLowerCase() != "/logs/" ? "├" : "└") + "─── CALLED: " + f.name);
     let promise: Promise<any> = f.apply(thisArg, args);
     if (req.url.toLowerCase() !== "/logs/") {
       promise.then((val) => {
+        Logger.doLog(val, true, thisArg, f, req, res, args);
         res.status(400).json(val);
       }).catch((reason) => {
-        res.status(200).json({ "error": "Promise rejected for: " + reason });
+        let error = { "error": "Promise rejected for: " + reason };
+        Logger.doLog(error, false, thisArg, f, req, res, args);
+        res.status(200).json(error);
       });
-    }
-  }
-}
-
-function printOnResponse(req: Request, res: Response, next: any) {
-  // var oldSend = res.send;
-  var oldJSON = res.json;
-
-  // res.send = function (data: any) {
-  //   if (data != undefined && req.url.trim().toLowerCase() != "/logs/") {
-  //     let str: string = JSON.stringify(data, null, 2);
-  //     responseBoxLog(str);
-  //   }
-  //   return oldSend.apply(res, arguments as any);
-  // }
-
-  res.json = function (data: any) {
-    if (data != undefined && req.url.trim().toLowerCase() != "/logs/") {
-      let str: string = JSON.stringify(data, null, 2);
-      responseBoxLog(str);
-    }
-    return oldJSON.apply(res, arguments as any);
-  }
-  next();
-}
-app.use(printOnResponse);
-
-let requestBoxLog = (msg: string) => {
-  let msgFin = "│ " + msg + " │";
-  let msgTop = "┌" + ((msgFinIn: string) => {
-    let str = "";
-    for (let i = 0; i < msgFinIn.length - 2; i++) {
-      str += "─";
-    }
-    return str;
-  })(msgFin) + "┐";
-  let msgBottom = "└" + ((msgFinIn: string) => {
-    let str = "";
-    for (let i = 0; i < msgFinIn.length - 2; i++) {
-      if (i == tabSpaceSize - 1) {
-        str += "┬";
-      } else {
-        str += "─";
-      }
-    }
-    return str;
-  })(msgFin) + "┘";
-  console.log(msgTop);
-  console.log(msgFin);
-  console.log(msgBottom);
-};
-
-let responseBoxLog = (msg: string) => {
-  let maxMsgLen = 60;
-  let msgFin = "";
-  let msgLen = Math.min(msg.length, maxMsgLen);
-  if (msg.length + 4 < maxMsgLen && !msg.includes("\n")) {
-    msgFin = "│ " + msg + ((value: string) => {
-      let concatStr = "";
-      for (let i = 0; i < msgLen - value.length - 4; i++) {
-        concatStr += " ";
-      }
-      return concatStr;
-    })(msg) + " │";
-  } else {
-    let strs = msg.split('\n');
-    if (strs.length <= 1) {
-      strs = [];
-      strs.unshift("RESPONSE OF:");
-      for (let i = 0; i < msg.length; i += maxMsgLen) {
-        strs.push(((msgFinIn: string) => {
-          let str = "│ ";
-          str += msg.substr(i, Math.min(i + maxMsgLen - 4, msgFinIn.length));
-          let fin = " │";
-          return str + fin;
-        })(msg));
-      }
     } else {
-      strs.unshift("RESPONSE OF:");
-      for (let i = 0; i < strs.length; i++) {
-        let line = strs[i];
-        if (line.length > maxMsgLen - 4) {
-          strs.splice(i, 1, line.substr(0, maxMsgLen - 4), line.substr(maxMsgLen - 4, line.length));
-        }
-        if (i > 50)
-          break
-      }
-      strs.forEach((val: any, ind: number) => {
-        strs[ind] = "│ " + val + ((value: string) => {
-          let concatStr = "";
-          for (let i = 0; i < msgLen - value.length - 4; i++) {
-            concatStr += " ";
-          }
-          return concatStr;
-        })(val) + " │";
-      });
+      Logger.doLog(null, true, thisArg, f, req, res, args);
     }
-    msgFin = strs.reduce((prev: string, curr: string, currInd: number, arr: string[]) => {
-      if (currInd == arr.length - 1) {
-        return prev += curr;
-      } else {
-        return prev += (curr + "\n");
-      }
-    }, "");
   }
-  let msgTop = "┌" + ((msgFinIn: string) => {
-    let str = "";
-    for (let i = 0; i < msgLen - 2; i++) {
-      if (i == tabSpaceSize - 1) {
-        str += "┴";
-      } else {
-        str += "─";
-      }
-    }
-    return str;
-  })(msg) + "┐";
-  let msgBottom = "└" + ((msgFinIn: string) => {
-    let str = "";
-    for (let i = 0; i < msgLen - 2; i++) {
-      str += "─";
-    }
-    return str;
-  })(msg) + "┘";
-  console.log(msgTop);
-  console.log(msgFin);
-  console.log(msgBottom);
-};
+}
 
 let getLogs = (dbv: any, reqv: Request, resv: Response) => {
   fs.readFile('logs.txt', { encoding: null, flag: undefined }, (err: any, data: Buffer) => {
